requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
pyyaml==6.0.1
